Code refactoring and check quality of the code
