+++ src/components/todo-form.tsx
@@ -1,21 +1,11 @@
 import React, { useRef, useState } from 'react';
 import AddTaskRoundedIcon from '@mui/icons-material/AddTaskRounded';
-import {
-  Box,
-  Button,
-  Container,
-  Dialog,
-  DialogActions,
-  DialogContent,
-  DialogTitle,
-  FormControl,
-  TextField,
-  Typography,
-} from '@mui/material';
+import { Box, Button, Container, FormControl, TextField } from '@mui/material';
 import { OnAddTodo } from '../types';
+import ErrorDialog from './error-dialog';
 
 const NewTodo = ({ handler }: OnAddTodo) => {
-  //Manage input fields
+  //Managing input fields
   const InputRef = useRef<HTMLInputElement>(null);
   const DesInputRef = useRef<HTMLInputElement>(null);
 
@@ -32,14 +22,14 @@ const NewTodo = ({ handler }: OnAddTodo) => {
   const onAddTodo = (event: React.FormEvent) => {
     event.preventDefault();
 
-    //Set current date to todos
+    //Setting current date to todos
     const currentDate = new Date();
 
-    //get title & description information
+    //Getting title & description information
     const enteredText = InputRef.current!.value;
     const enteredDescription = DesInputRef.current!.value;
 
-    //Manage the errors
+    //Managing the errors
     switch (true) {
       case enteredText === '' || enteredDescription === '':
         setErrorMessage('Please fill out both of the fields!');
@@ -70,7 +60,7 @@ const NewTodo = ({ handler }: OnAddTodo) => {
       maxWidth='lg'
       sx={{ mt: 2 }}
     >
-      {/* todo-form */}
+      {/* todoform */}
       <Box sx={{ width: '35rem' }}>
         <form onSubmit={onAddTodo}>
           <Box sx={{ mb: 2 }}>
@@ -111,28 +101,11 @@ const NewTodo = ({ handler }: OnAddTodo) => {
       </Box>
 
       {/* Error handling */}
-      <Dialog
+      <ErrorDialog
         open={open}
+        message={errorMessage}
         onClose={handleClose}
-        disableEnforceFocus
-        closeAfterTransition={false}
-        aria-labelledby='alert-dialog-title'
-        aria-describedby='alert-dialog-description'
-      >
-        <DialogTitle id='alert-dialog-title'>Error</DialogTitle>
-        <DialogContent>
-          <Typography id='alert-dialog-description'>{errorMessage}</Typography>
-        </DialogContent>
-        <DialogActions>
-          <Button
-            aria-label='this is area label'
-            onClick={handleClose}
-            autoFocus
-          >
-            Ok
-          </Button>
-        </DialogActions>
-      </Dialog>
+      />
     </Container>
   );
 };
